requirements.txt
fire==0.1.3
riemann-tx==2.0.0
virtualenv==16.6.1
